is_type_obvious(phi) = True

+++ tests/end_to_end/float-conversions/irix-o2-out.c
@@ -13,13 +13,13 @@ void *test(void)
     {
         phi_f18 = temp_f18 + 4294967296.0;
     }
-    D_410238 = (f64) phi_f18;
+    D_410238 = phi_f18;
     temp_f8 = (f32) D_410240;
     phi_f8 = temp_f8;
     if (D_410240 < 0)
     {
         phi_f8 = temp_f8 + 4294967296.0f;
     }
-    D_410230 = (f32) phi_f8;
+    D_410230 = phi_f8;
     return &D_410240;
 }
