+++ tests/end_to_end/multiple-assigns/irix-g-out.c
@@ -8,7 +8,7 @@ s32 test(s32 arg0)
     if (arg0 == 5)
     {
 loop_1:
-        D_410150 = (s32) phi_a0;
+        D_410150 = phi_a0;
         temp_a0 = phi_a0 + 1;
         D_410150 = temp_a0;
         temp_a0 = temp_a0 + 1;
